requirements.txt
pandas
playwright
python-dotenv
requests